Move routers to server package.
