+++ internal/server/interface.go
@@ -1,4 +1,4 @@
-package routers
+package server
 
 import (
 	"github.com/gin-gonic/gin"
