+++ internal/server/short.go
@@ -1,4 +1,4 @@
-package routers
+package server
 
 import (
 	"fmt"
